refactor: Convert SearchQA and SerpAPILoader from JS to TS
