+++ api/app/clients/tools/manifest.json
@@ -89,7 +89,7 @@
     "name": "Search and QA",
     "pluginKey": "search-qa",
     "description": "Useful for when you need to answer questions about current events",
-    "icon": "https://upload.wikimedia.org/wikipedia/en/e/e5/FinnAdventureTime.png",
+    "icon": "https://i.scdn.co/image/ab6761610000e5ebc94fb92f8143c3637c6f7b80",
     "isAuthRequired": "false",
     "authConfig": []
   },
